New link added with valuable information
The link I added includes a lot of general information about major depression that can be implemented/added into the project if we haven't done so already. I like the link in the website that allows you to ask a doctor questions. Maybe we can add a referral to a licensed professional or clinic.