Logic for Quiz Scores
The depression screening test will now correctly categorize the severity of depression based on the user's score. This score will be stored in the user's account onto our DynamoDB Table which we will be used for enhanced functionalities. The score and severity correlation may also be referenced for further functionality of Good Vibes.